requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
gunicorn==21.2.0
